# SurveyTool/LearningPlatform/mongodb-api-documentation-swagger.yaml
...
swagger: '2.0'
info:
  version: v1
  title: ResponsiveInsight API
  description: |
    This API will be used both internally for the application, and externally for integration. 
    
    Each path below will have ETAGs. The /surveys/ and /surveydefintions/ will have different ETAGS. 
host: api-surveytool.orientsoftware.asia
basePath: /api
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
...
parameters:
  surveyId:
    name: surveyId
    in: path
    description: The survey identifier
    type: string
    required: true
...
  reportId:
    name: reportId
    in: path
    description: the report identifier
    type: integer
    required: true
...
  themeId:
    name: themeId
    in: path
    description: the theme identifier
    type: string
    required: true
paths:
...
  /surveys/{surveyId}/:
    parameters:
      - $ref: '#/parameters/surveyId'
    get:
      description: Get the shallow survey object (not including survey settings). It will include the surveydefinition etag. The polling for changes should call this resource. If modified, the code will check if the surveydefinition etag matches the survey definition etag in the model. If not match, the code will fetch the survey definition. 
      tags: [surveys]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    put:
      description: update survey.
      tags: [surveys]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /surveys/{surveyId}/delete/:
    parameters:
      - $ref: '#/parameters/surveyId'
    put:
      description: soft delete a survey follow id.
      tags: [surveys]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /surveys/{surveyId}/restore/:
    parameters:
      - $ref: '#/parameters/surveyId'
    put:
      description: restore a survey follow id.
      tags: [surveys]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /surveys/{surveyId}/publish/:
    parameters:
      - $ref: '#/parameters/surveyId'
    post:
      description: publish a survey follow id.
      tags: [surveys]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /surveys/{surveyId}/file/:
    parameters:
      - $ref: '#/parameters/surveyId'
    get:
      tags: [surveys]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /surveys/{surveyId}/export/:
    parameters:
      - $ref: '#/parameters/surveyId'
    post:
      description: export a survey to json file follow id.
      tags: [surveys]
      produces:
        - application/json
      responses:
        '200':
          description: OK
...
  /surveys/{surveyId}/duplicate/:
    parameters:
      - $ref: '#/parameters/surveyId'
    get:
      description: duplicate a survey follow id.
      tags: [surveys]
      produces:
        - application/json
      responses:
        '200':
          description: OK
...
  /themes/{themeId}/survey/:
    parameters:
      - $ref: '#/parameters/themeId'
    get:
      tags: [layouts and themes]
      produces:
        - application/json
      responses:
        '200':
          description: OK
...
  /reports/{reportId}/definition/:
    parameters:
      - $ref: '#/parameters/reportId'
    get:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    put:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    post:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
...